Ability to move actors around the grid.
Made an interface for input helpers and added a new input helper to figure out if a button was pressed and does not allow repeating. Moved input-related classes to their own package.

Added an actor class that only has a location for now. Added code in the Grid to sync Actor location with Tile location.

Added some game logic to allow moving actors around. Press enter to select an actor, then enter to select an empty passable square. There is no visual indication that an actor is selected yet.