Working basic menu.
I think this is only about 75% horrible. Any better architecture suggestions are welcome.
I also want to make a "message" class or something. I think the left portion of the bottom should be for menus, and the right portion of the bottom (maybe as much as 75%) should be for messages about what's going on.
Someone's going to need to clean up the API on Grid.java soon.