Making the screen resizeable and keeping static constants private, passing around width and height instead. The screen must be at least 1024x768, but can be arbitrarily larger and will stretch the grid and information area to fit. The grid doesn't scroll or anything fancy yet.
Before the resize feature can be considered complete, we need to modify the bottom section to not waste so much room when expanding the screen (i.e. give it a static vertical size in pixels since we can't make the screen smaller anyway) and modify the grid to take advantage of the space it has or cope with the space it doesn't have.