Working attacks and damage.
I haven't tested with higher numbers, but this damage calculation is similar to what LoL uses, so it should scale well. This isn't where I want the damage code ultimately, but we'll deal with that once we add more than one kind of attack.

The strength of the attacking hero is randomized slightly, so the base damage of the attack between 90% and 140% of the attacker's strength.

The target's defense is then taken into account, so the target takes a reduced damage percent. The damage taken is multiplied by 100 / (100 + defense). For example, if the target has 10 defense, the target would only take 90.9% damage. If the target has 50 defense, he would take 66% damage. If the target has 150 defense, he would take 40% damage. It never completely mitigates all the damage, but like proven in all the LoL mathcraft, additional defense adds linearly to the effective hp of the defender.