Basics of equipment.
I'm not sure how much I like making one class per equipment type. I might make generic classes for types of equipment (weapon, shield, etc.) and make factories that generate weapons of certain enumeration types. Or something completely different. But this works for now, so it's good enough to get us to the next step.

We need to add code to unequip weapons from certain slots and to specify which hand to replace when equipping something to the hand slot. Right now, it always replaces the left hand if the right hand is already holding something, which is less than ideal.