More advanced targeting mechanics
Now you can target yourself with attacks, and attacks automatically affect everything within the current area of effect specified for the Grid.java

The next step is to make applying the effect happen in its own class so that we can move that logic out of the Game class and into somewhere more generic. Basic attacks will be converted to an effect like that, and adding spells will be as simple as adding a new effect.