Making AoE code more robust
And added a new (tested and working) LineAOE. Moved a lot of code from the abstract class to the specific class. Each class should take care of its own code, because in the case of line attacks, the relative positions of affected squares could vary based on the position of the reference square.