Cure spell.
Also improvements to the Effect abstract class so that more things can be handled automatically when creating a new effect. I just made placeholder changes to the menu to test out the cure spell. The next step is to make an actual magic menu where you can choose your spell from a list.

I also may modify the move command to be an Effect, since that class is generic enough to handle it.